Update README to describe Atree collision handler more
This commit has 2 changes:

- Update list of fxamacker/cbor users.

- Update description of novel hash collision handler:

This new hash collision handling method balances speed, security, storage size, etc.  It uses a fast noncryptographic 64-bit hash and if there is a 64-bit hash collision, it uses deferred and segmented 256-bit cryptographic digest (in 64-bit segments to save space).  If there is a collision on the entire combined hash digest, the collision is handled gracefully (not catastrophic).  By default, it uses [CircleHash64f](https://github.com/fxamacker/circlehash) and BLAKE3 for a max combined digest size limit of 320 bits.  Larger digest sizes are possible by replacing BLAKE3 with SHA3-512, etc. for a max digest size limit of 576 bits.  If the max digest size limit is too large, it can be truncated by increments of 64-bits until desired tradeoffs (e.g., collision resistance) are reached.  By design, the cryptographic hash is only computed if there is a collision on the seeded 64-bit fast hash and it is only computed once.